+++ arklight/cli/templates/production.py
@@ -50,7 +50,7 @@ def build(name: str) -> dict[str, str]:
 
 
 _SITE_PY = '''\
-from arklight import *
+# include <stdlib.ARKlight>
 
 from pages.about import about
 from pages.home import home